# lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Repody",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/uploads/capabilities": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "Upload Capabilities",
        "operationId": "upload_capabilities_v1_uploads_capabilities_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Capabilities V1 Uploads Capabilities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads/presign": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Presign Uploads",
        "operationId": "presign_uploads_v1_uploads_presign_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads/confirm": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Confirm Uploads",
        "operationId": "confirm_uploads_v1_uploads_confirm_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Upload Files",
        "operationId": "upload_files_v1_uploads_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_files_v1_uploads_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_files_v1_uploads_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_files_v1_uploads_post"
      },
...
      "ConfirmUploadItem": {
        "properties": {
          "storageKey": {
            "type": "string",
            "title": "Storagekey"
          },
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "storageKey",
          "fileName",
          "mimeType",
          "size"
        ],
        "title": "ConfirmUploadItem"
      },
      "ConfirmUploadRequest": {
        "properties": {
          "storageKeys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Storagekeys"
          }
        },
        "type": "object",
        "required": [
          "storageKeys"
        ],
        "title": "ConfirmUploadRequest"
      },
      "ConfirmUploadResponse": {
        "properties": {
          "uploads": {
            "items": {
              "$ref": "#/components/schemas/ConfirmUploadItem"
            },
            "type": "array",
            "title": "Uploads"
          }
        },
        "type": "object",
        "required": [
          "uploads"
        ],
        "title": "ConfirmUploadResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PresignFileRequest": {
        "properties": {
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "documentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentid"
          }
        },
        "type": "object",
        "required": [
          "fileName",
          "mimeType",
          "size"
        ],
        "title": "PresignFileRequest"
      },
      "PresignRequest": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/PresignFileRequest"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "PresignRequest"
      },
      "PresignResponse": {
        "properties": {
          "uploadMode": {
            "type": "string",
            "title": "Uploadmode",
            "description": "presigned | api"
          },
          "uploads": {
            "items": {
              "$ref": "#/components/schemas/PresignedUploadItem"
            },
            "type": "array",
            "title": "Uploads"
          }
        },
        "type": "object",
        "required": [
          "uploadMode",
          "uploads"
        ],
        "title": "PresignResponse"
      },
      "PresignedUploadItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "storageKey": {
            "type": "string",
            "title": "Storagekey"
          },
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "documentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentid"
          },
          "uploadUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadurl"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": "PUT"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "storageKey",
          "fileName",
          "mimeType",
          "size"
        ],
        "title": "PresignedUploadItem"
      },
...
      "UploadItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "storageKey": {
            "type": "string",
            "title": "Storagekey"
          },
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "id",
          "storageKey",
          "fileName",
          "mimeType",
          "size"
        ],
        "title": "UploadItem"
      },
      "UploadResponse": {
        "properties": {
          "uploads": {
            "items": {
              "$ref": "#/components/schemas/UploadItem"
            },
            "type": "array",
            "title": "Uploads"
          }
        },
        "type": "object",
        "required": [
          "uploads"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}